refactor(cli): group pipeline stages into pipeline/ folder
Flatten `src/` no longer reflected the workflow: io / csv_input /
structure / transform / output were sorted alphabetically next to
cross-cutting modules like args / domain / error, hiding the fact that
they are the six stages of a single pipeline.

Reorganize so the pipeline's sub-modules live under `pipeline/`:

    src/
    ├── main.rs, lib.rs
    ├── args.rs
    ├── error.rs
    ├── domain.rs
    └── pipeline/
        ├── mod.rs        # orchestration (FileTask, run, process_file)
        ├── io.rs         # read/write
        ├── csv_input.rs  # Stage 2: parse
        ├── structure.rs  # Stage 3: structure
        ├── transform.rs  # Stage 4: aggregate
        └── output.rs     # Stage 4/5: project + serialize helpers

`ls src/pipeline/` now shows the six files of the pipeline grouped
together, while the crate-wide concerns (args, error, domain) remain
flat at the top.

Module paths updated:
- `crate::csv_input` → `crate::pipeline::csv_input` (similar for the
  others)
- Sibling references inside `pipeline/` use `super::` or
  `crate::pipeline::` as appropriate
- `for_testing` re-exports updated accordingly

No behavior change; 62 tests pass.
