+++ cli/cli/src/pipeline/output.rs
